massive refactor with better separation
