+++ back_testing/sma_cross_backtest.py
@@ -17,60 +17,7 @@ from algo_trading.config.controllers import (
 from algo_trading.utils.utils import str_to_dt
 from algo_trading.strategies.events import TradeEvent
 
-from dags.sma_cross_dag import update_redis
-
 from controllers import TestPeriodController
-
-
-def cross_up(data: pd.DataFrame, index: int) -> bool:
-    """Checks to see if a cross up occured by looking at
-    the current date 7 and 21 day SMA and the previous date
-    7 and 21 day SMA. Finally, we only consider cross up
-    when the close price > 50 ay SMA otherwise the market
-    is considered bearish.
-
-    Args:
-        data (pd.DataFrame): Data to parse SMA info.
-        index (int): Indicates current day. index + 1 = prev day.
-
-    Returns:
-        bool: True if all conditions are met.
-    """
-    return (
-        (
-            data[ColumnController.ma_7.value].iloc[index]
-            >= data[ColumnController.ma_21.value].iloc[index]
-        )
-        and (
-            data[ColumnController.ma_7.value].iloc[index - 1]
-            < data[ColumnController.ma_21.value].iloc[index - 1]
-        )
-        # and (
-        #     data[ColumnController.close.value].iloc[index]
-        #     > data[ColumnController.ma_50.value].iloc[index]
-        # )
-    )
-
-
-def cross_down(data: pd.DataFrame, index: int) -> bool:
-    """Checks to see if a cross down occured by looking at
-    the current date 7 and 21 day SMA and the previous date
-    7 and 21 day SMA.
-
-    Args:
-        data (pd.DataFrame): Data to parse SMA info.
-        index (int): Indicates current day. index + 1 = prev day.
-
-    Returns:
-        bool: True if all conditions are met.
-    """
-    return (
-        data[ColumnController.ma_7.value].iloc[index]
-        < data[ColumnController.ma_21.value].iloc[index]
-    ) and (
-        data[ColumnController.ma_7.value].iloc[index - 1]
-        >= data[ColumnController.ma_21.value].iloc[index - 1]
-    )
 
 
 class SMACrossBackTester:
@@ -163,14 +110,15 @@ class SMACrossBackTester:
         except AttributeError:
             if self.period:
                 if self.period == TestPeriodController.max:
-                    data = self.db_repo.get_all(self.ticker)
+                    self._price_data = self.db_repo.get_all(self.ticker)
                 else:
-                    data = self.db_repo.get_days_back(self.ticker, self.days_back)
+                    self._price_data = self.db_repo.get_days_back(
+                        self.ticker, self.days_back
+                    )
             elif self.start_date:
-                data = self.db_repo.get_since_date(self.ticker, self.start_date)
-            self._price_data = data.sort_values(
-                [ColumnController.date.value], ascending=True
-            ).reset_index(drop=True)
+                self._price_data = self.db_repo.get_since_date(
+                    self.ticker, self.start_date
+                )
             return self._price_data
 
     def _init_fake_key_value(self) -> Dict:
@@ -260,38 +208,45 @@ class SMACrossBackTester:
 
         sma = SMACross(self.ticker, fake_db_repo, fake_kv_repo)
         starting_cap = self.capital
-        num_shares = 0
         num_trades = 0
         if init_status == StockStatusController.buy.value:
             num_shares = self._get_num_shares(
                 self.capital, self.price_data[ColumnController.close.value].iloc[0]
             )
+            init_message = f"{num_shares} shares"
+        else:
+            num_shares = 0
+            init_message = f"${self.capital}"
+
         print("\n*******************************************************")
         print(
-            f"Beginning SMA Cross strategy with ${self.capital} on "
+            f"Beginning SMA Cross strategy with {init_message} at price "
+            + f"{self.price_data[ColumnController.close.value].iloc[0]} on "
             + f"{self.price_data[ColumnController.date.value].iloc[0]}"
         )
         print("*******************************************************\n")
 
         for idx, row in self.price_data.iterrows():
             if idx == 0:
+                # Skip the first day.
                 continue
             else:
 
                 # Current key/val store for self.ticker
                 curr = json.loads(fake_kv_repo.handler.get(self.ticker))
 
-                if cross_up(self.price_data[: (idx + 1)], idx):
+                if SMACross.cross_up(self.price_data[: (idx + 1)], idx):
                     curr[ColumnController.last_cross_up.value] = dt_to_str(
                         self.price_data[ColumnController.date.value].iloc[idx]
                     )
                     fake_kv_repo.handler.set(self.ticker, curr)
 
-                elif cross_down(self.price_data[: (idx + 1)], idx):
+                elif SMACross.cross_down(self.price_data[: (idx + 1)], idx):
                     # Checks the case when we had a cross up in bear market
                     if str_to_dt(
                         curr[ColumnController.last_cross_down.value]
                     ) < str_to_dt(curr[ColumnController.last_cross_up.value]):
+
                         curr[ColumnController.last_cross_down.value] = dt_to_str(
                             self.price_data[ColumnController.date.value].iloc[idx]
                         )
@@ -324,22 +279,28 @@ class SMACrossBackTester:
                 num_shares, self.price_data.iloc[-1][ColumnController.close.value]
             )
             print(
-                f"Sold {num_shares} shares at price "
+                f"\nFinished with {num_shares} shares at price "
                 + f"{self.price_data.iloc[-1][ColumnController.close.value]} "
-                + f"on {self.price_data.iloc[-1][ColumnController.date.value]} "
-                + f"for a new capital of {self.capital}."
+                + f"on {self.price_data.iloc[-1][ColumnController.date.value]}."
             )
+            print(f"Selling all for a new capital of {self.capital}.")
         percent_change = self._get_percent_change(starting_cap, self.capital)
-        print(f"Starting cap: {starting_cap}, final cap: {self.capital}.")
-        print(f"Change over {self.period.value} is {percent_change} %.")
+        print(f"Starting cap: {starting_cap}")
+        print(f"Final cap: {self.capital}")
+        print(f"Change over {self.period or self.start_date} is {percent_change} %.")
         print(f"Number of trades: {num_trades}")
 
 
 if __name__ == "__main__":
-    from algo_trading.constants import DB_INFO
+    from algo_trading.config import DB_INFO
 
     ticker = "aapl"
     tester = SMACrossBackTester(
-        ticker, DB_INFO, DBHandlerController.postgres, "max", capital=1000
+        ticker=ticker,
+        db_info=DB_INFO,
+        db_handler=DBHandlerController.postgres,
+        # period="3mo",
+        start_date="2020-01-01",
+        capital=1000,
     )
     tester.test()
