+++ docker-compose.yml
@@ -12,7 +12,7 @@ services:
     volumes:
       - postgres_data:/var/lib/postgresql/data/pgdata
     env_file:
-      - local.env
+      - .env
     ports:
       - 5432:5432
     networks:
@@ -24,7 +24,7 @@ services:
     depends_on:
       - postgres
     env_file:
-      - local.env
+      - .env
     ports:
       - 80:80
     networks:
