+++ sample.env
@@ -1,6 +1,6 @@
 POSTGRES_USER=dev
 POSTGRES_PASSWORD=password
-POSTGRES_DB=dev
+POSTGRES_DB=price_db
 POSTGRES_HOST=localhost
 POSTGRES_PORT=5432
 PGDATA=/var/lib/postgresql/data/pgdata
